requirements.txt
Flask[async]
Flask-CORS
pymongo
openai
python-dotenv
requests
aiohttp
motor
aiohttp
pandas
openpyxl
hypercorn
quart
rich
uvicorn
